third_party/WebKit/LayoutTests/fast/events/wheelevent-in-scrolling-div-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS event.deltaY is window.expectedScrollTop
PASS event.deltaX is window.expectedScrollLeft
PASS event.wheelDeltaY is window.expectedScrollTop*-120
PASS event.wheelDeltaX is window.expectedScrollLeft*-120
PASS event.wheelDelta is window.expectedScrollTop*-120
PASS div.scrollTop is window.expectedScrollTop
PASS div.scrollLeft is window.expectedScrollLeft
